agent-memory/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-memory"
version = "1.0.0"
description = "Persistent memory service for multi-agent systems"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "aiosqlite>=0.20.0",
    "python-ulid>=2.2.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
agent-memory = "agent_memory.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_memory"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/agent_memory"]
omit = ["tests/*"]
